Compilable

+++ includes/global/utils/utils.h
@@ -6,7 +6,7 @@ void ncurses_init();
 void ncurses_stop();
 void ncurses_colors();
 void ncurses_init_mouse();
-int mouse_getpos(int *x, int *y);
+int mouse_getpos(short *x, short *y);
 void palette();
 void center_string(const char* str, const int dest_len, char* dest);
